# openapi.yaml
openapi: 3.1.0
info:
  title: Colophon
  version: 1.0.0
  description: |
    Two public-domain editions, served from Cloudflare Workers: the American
    Standard Version (1901) and the Douay-Rheims. Both corpora are embedded in
    the Worker, so no request performs any I/O.

    Every response names the translation it came from, and carries the edition,
    revision and generation identifiers of that text. Pass ?translation= to
    choose; an unknown id is refused rather than silently answered from the
    default.

    Versification belongs to the edition, not to the canon. Of the 66 books the
    two share, 32 are numbered differently — so a verse coordinate is only
    meaningful paired with a translation.

    Canon awareness: every book carries its position in the Protestant (66),
    Catholic (73), and Orthodox Greek (76) canons. The ASV contains no
    deuterocanonical text, so those books are metadata only there; the
    Douay-Rheims supplies seven of them.
  license:
    name: Public Domain
    url: https://ebible.org/asv/
servers:
  - url: https://bible-api.workers.dev
paths:
...
  /books:
    get:
      summary: List books in a canon's order
      parameters:
        - name: translation
          in: query
          required: false
          description: |
            Which translation to read. Defaults to `asv`.

            An unknown id is a 404 naming what this deployment does serve, never
            a silent fall back to the default: a caller asking for a text this
            deployment does not hold must be told, not handed a different Bible.

            Verse coordinates belong to a translation, not to the canon. Editions
            disagree about versification — of the 66 books the ASV and the
            Douay-Rheims share, 32 differ, Esther and Daniel by whole chapters —
            so a reference is only meaningful paired with the translation it was
            resolved against.
          schema: { type: string, default: asv, example: asv }
        - name: tradition
          in: query
          required: false
          schema:
            type: string
            enum: [protestant, catholic, orthodox_greek]
            default: protestant
      responses:
        "200":
          description: Books ordered by their position in the requested tradition.
          content:
            application/json:
              schema:
                type: object
                required: [tradition, translation, books]
                additionalProperties: false
                properties:
                  tradition: { type: string, enum: [protestant, catholic, orthodox_greek] }
                  translation: { $ref: "#/components/schemas/Translation" }
                  books:
                    type: array
                    items: { $ref: "#/components/schemas/Book" }
        "400": { $ref: "#/components/responses/BadRequest" }
  /books/{id}:
    get:
      summary: A single book
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, example: JHN }
      responses:
        "200":
          description: Book metadata.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Book" }
        "404": { $ref: "#/components/responses/NotFound" }
  /books/{id}/chapters/{num}:
    get:
      summary: One chapter, in full
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, example: PSA }
        - name: num
          in: path
          required: true
          schema: { type: integer, minimum: 1, example: 23 }
      responses:
        "200":
          description: Every verse of the chapter.
          content:
            application/json:
              schema:
                type: object
                required: [id, book_id, chapter, descriptive_title, subscription, verses]
                additionalProperties: false
                properties:
                  id: { type: string, example: PSA.23 }
                  book_id: { type: string, example: PSA }
                  chapter: { type: integer }
                  descriptive_title:
                    type: [string, "null"]
                    description: >
                      The superscription printed above the chapter, as on 116 of
                      the Psalms ("A Psalm of David."). Null elsewhere.
                  subscription:
                    type: [string, "null"]
                    description: >
                      A line printed below the chapter's last verse. Habakkuk 3
                      is the only chapter in the ASV with one. The source marks
                      it with the same element as a superscription, so the two
                      are told apart by position, not markup.
                  verses:
                    type: array
                    items: { $ref: "#/components/schemas/ChapterVerse" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
  /passages:
    get:
      summary: Resolve a natural-language reference
      description: |
        Accepts references such as `John 3:16`, `1 Cor 13:4-7,13`,
        `Genesis 1:1-2:3`, `Jude 5`, or a bare book name. Ambiguous
        abbreviations are rejected: `Jo` could mean Job, Joel, John, Jonah, or
        Joshua.

        A one-chapter book takes either form: `Jude 5` and `Jude 1:5` both
        resolve, as do `2 John 10` and `2 John 1:10`.

        Errors divide by cause. Input that cannot be understood is a 400;
        a reference that is well formed but names something absent — an unknown
        book, or a chapter or verse past the end — is a 404, matching what the
        `/books` endpoints return for the same condition.
      parameters:
        - name: translation
          in: query
          required: false
          description: |
            Which translation to read. Defaults to `asv`.

            An unknown id is a 404 naming what this deployment does serve, never
            a silent fall back to the default: a caller asking for a text this
            deployment does not hold must be told, not handed a different Bible.

            Verse coordinates belong to a translation, not to the canon. Editions
            disagree about versification — of the 66 books the ASV and the
            Douay-Rheims share, 32 differ, Esther and Daniel by whole chapters —
            so a reference is only meaningful paired with the translation it was
            resolved against.
          schema: { type: string, default: asv, example: asv }
        - name: ref
          in: query
          required: true
          description: |
            At most 512 characters, and at most 64 comma-separated parts. Both
            are far past any real citation; they exist so that a reference which
            was never going to be served is refused before it is parsed.
          schema: { type: string, maxLength: 512, example: "John 3:16" }
        - name: numbering
          in: query
          required: false
          description: |
            Psalms only; every other book is numbered alike in all traditions.

            `hebrew` follows the Masoretic text, where a psalm's superscription
            is verse 1, so every English verse shifts up by one on the 116 Psalms
            that carry one. The superscription itself is returned as verse 0.

            `greek` follows the Septuagint and Vulgate, as Catholic and Orthodox
            sources cite them. This changes the psalm number as well as the
            verses, and not by a constant: Greek 9 is Hebrew 9 and 10 together,
            Greek 113 is Hebrew 114 and 115, while Hebrew 116 and 147 each split
            in two. So Greek 50 is the Miserere, printed here as Psalm 51.
          schema:
            type: string
            enum: [english, hebrew, greek]
            default: english
      responses:
        "200":
          description: The verses the reference selects.
          content:
            application/json:
              schema:
                type: object
                required: [reference, translation, numbering, verses]
                additionalProperties: false
                properties:
                  reference:
                    type: string
                    description: The reference normalized for display.
                    example: 1 Corinthians 13:4-7,13
                  translation: { $ref: "#/components/schemas/Translation" }
                  numbering: { type: string, enum: [english, hebrew, greek] }
                  verses:
                    type: array
                    items: { $ref: "#/components/schemas/PassageVerse" }
        "400":
          description: >
            The reference could not be understood: malformed syntax, or an
            abbreviation ambiguous enough to match several books.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "404":
          description: >
            The reference was understood but names something absent: an unknown
            book, a deuterocanonical book the ASV does not contain, or a chapter
            or verse beyond the end of the book.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "413":
          description: The reference covers more than 500 verses.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /search:
    get:
      summary: Ranked full-text search
      description: |
        BM25 ranking over an inverted index of all 31,102 verses.

        A verse must contain **every** term to match, so `total` is the number of
        verses containing all of them — not the number containing any.

        Each term matches more than its literal spelling. Words are grouped into
        morphological families, so `speak` also finds `spake` and `spoken`, and
        `say` also finds `said` and `saith` — forms no prefix rule could reach,
        and among the most common words in this translation. The final term
        additionally matches by prefix, so `believet` finds `believeth`.

        Ranking: every result contains every term, so BM25's length
        normalization orders them, putting the most concise matching verse first.
      parameters:
        - name: translation
          in: query
          required: false
          description: |
            Which translation to read. Defaults to `asv`.

            An unknown id is a 404 naming what this deployment does serve, never
            a silent fall back to the default: a caller asking for a text this
            deployment does not hold must be told, not handed a different Bible.

            Verse coordinates belong to a translation, not to the canon. Editions
            disagree about versification — of the 66 books the ASV and the
            Douay-Rheims share, 32 differ, Esther and Daniel by whole chapters —
            so a reference is only meaningful paired with the translation it was
            resolved against.
          schema: { type: string, default: asv, example: asv }
        - name: q
          in: query
          required: true
          description: |
            Between 2 and 512 characters, resolving to at most 12 distinct terms;
            duplicates do not count against that, so a repeated word is one term.

            Twelve is past any real phrase, and no query is refused for being
            made of common words: `and it came to pass in the days of the king`
            is served like any other.
          schema: { type: string, minLength: 2, maxLength: 512, example: good shepherd }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
        - name: offset
          in: query
          schema: { type: integer, minimum: 0, default: 0 }
      responses:
        "501":
          description: |
            The full-text index covers the ASV only. Rather than answer from it
            while labelling the response with another translation, this route
            declines and names what it does cover.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "200":
          description: Ranked matches, best first.
          content:
            application/json:
              schema:
                type: object
                required: [query, total, limit, offset, truncated, translation, results]
                additionalProperties: false
                properties:
                  query: { type: string }
                  total: { type: integer }
                  limit: { type: integer }
                  offset: { type: integer }
                  truncated:
                    type: boolean
                    description: >
                      True when the query's most selective term was itself common
                      enough to hit the scan cap, leaving `total` a lower bound.
                      Only single very common words normally reach this; adding a
                      second term makes the query exact again.
                  translation: { $ref: "#/components/schemas/Translation" }
                  results:
                    type: array
                    items: { $ref: "#/components/schemas/SearchResult" }
        "400": { $ref: "#/components/responses/BadRequest" }
components:
  responses:
    BadRequest:
      description: The request was malformed; `detail` says how.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: No such resource.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
    Error:
      type: object
      required: [error, detail]
      additionalProperties: false
      properties:
        error:
          type: string
          enum: [bad_request, not_found, payload_too_large, internal_error]
        detail: { type: string }
    Translation:
      type: object
      required: [id, name, language, license, year]
      additionalProperties: false
      properties:
        id: { type: string, enum: [asv, dra] }
        name: { type: string }
        language: { type: string }
        license: { type: string }
        year: { type: integer }
    Book:
      type: object
      required:
        [id, name, testament, is_deuterocanon, chapters_count, author, genre,
         approximate_date, data_availability, canons]
      additionalProperties: false
      properties:
        id: { type: string, example: JHN }
        name: { type: string, example: John }
        testament: { type: string, enum: [OT, NT] }
        is_deuterocanon: { type: boolean }
        chapters_count:
          type: integer
          description: Zero for deuterocanonical books, which carry no text here.
        author:
          type: string
          description: >
            The traditional ascription, not a critical judgement — "Moses" for the
            Pentateuch, "Ezra" for Chronicles. Catalogue metadata.
        genre: { type: string }
        approximate_date:
          type: string
          description: >
            A broad range spanning the composition debate rather than settling it.
        data_availability:
          type: string
          enum: [full, metadata_only]
        canons:
          type: object
          required: [protestant, catholic, orthodox_greek]
          additionalProperties: false
          description: 1-based position in each canon, or null if absent from it.
          properties:
            protestant: { type: [integer, "null"] }
            catholic: { type: [integer, "null"] }
            orthodox_greek: { type: [integer, "null"] }
    ChapterVerse:
      type: object
      required: [id, number, text, hebrew_numbering, note]
      additionalProperties: false
      properties:
        id: { type: string, example: PSA.23.1 }
        number: { type: integer }
        text: { type: string }
        hebrew_numbering:
          type: integer
          description: >
            The verse's number in Hebrew reckoning. Differs from `number` only
            on the 116 Psalms carrying a superscription.
        note:
          type: [string, "null"]
          description: >
            Present on the 16 verses the ASV omits, which appear as numbered
            placeholders with empty text and this explanatory footnote.
    PassageVerse:
      type: object
      required: [id, book, chapter, verse, text, note]
      additionalProperties: false
      properties:
        id: { type: string, example: JHN.3.16 }
        book: { type: string, example: JHN }
        chapter: { type: integer }
        verse:
          type: integer
          description: >
            Zero denotes a Psalm superscription, returned when Hebrew numbering
            selects verse 1 of a titled Psalm.
        text: { type: string }
        note: { type: [string, "null"] }
    SearchResult:
      type: object
      required: [id, book, chapter, verse, text, score]
      additionalProperties: false
      properties:
        id: { type: string }
        book: { type: string }
        chapter: { type: integer }
        verse: { type: integer }
        text: { type: string }
        score: { type: number, description: BM25 relevance; higher is better. }